refactor: global 패키지 import 변경 #131
